requirements.txt
fastapi
kubernetes
prometheus-client
pydantic
pyjwt
python-dotenv
uvicorn
